V1.2
Added a delete sprite in another place because I noticed it got stuck on the screen once.
Added a switch to set the receiver into permanent receiver mode so it couldn't accidentally be put in Transimitter mode.
Changed the delay on showing the # of times fed so it would only happen once an hour.  At once a minute it was killing the battery within 24 hours.